remote/vision-service/pyproject.toml
[project]
name = "a22-vision-service"
version = "0.1.0"
description = "Remote vision preprocessing service for A22"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.30.0",
  "pydantic==2.7.0",
  "Pillow==10.4.0",
  "torch==2.10.0",
  "nvidia-cublas-cu12==12.8.4.1",
  "nvidia-cuda-cupti-cu12==12.8.90",
  "nvidia-cuda-nvrtc-cu12==12.8.93",
  "nvidia-cuda-runtime-cu12==12.8.90",
  "nvidia-cudnn-cu12==9.10.2.21",
  "nvidia-cufft-cu12==11.3.3.83",
  "nvidia-curand-cu12==10.3.9.90",
  "nvidia-cusolver-cu12==11.7.3.90",
  "nvidia-cusparse-cu12==12.5.8.93",
  "nvidia-cusparselt-cu12==0.7.1",
  "nvidia-nccl-cu12==2.27.5",
  "nvidia-nvjitlink-cu12==12.8.93",
  "nvidia-nvtx-cu12==12.8.90",
  "triton==3.6.0",
  "transformers==4.57.6",
  "accelerate==1.3.0",
  "safetensors==0.7.0",
]

[tool.uv]
package = false
